Немного кастрации физики и новый domique
