+++ src/ru/muwa/shq/objects/GameObject.java
@@ -11,7 +11,8 @@ public abstract class GameObject
     /**
      * Поля
      */
-    public static final String IMG_PATH = "C:\\Users\\darkm\\IdeaProjects\\shq_rdr\\src\\ru\\muwa\\shq\\textures\\";
+
+    public static final String IMG_PATH = "D:\\IdeaProjects\\src\\ru\\muwa\\shq\\textures\\";
     //TODO: Вован сделай такое же поле для себя, а мое закомментируй при работе в своей ветке.
     //public static final String IMG_PATH = "C:\\sosok\\shq_rdr\\src\\ru\\muwa\\shq\\textures\\";
     protected int x, y; //Координаты положения в пространстве
